hajjinfo/requirements.txt
asgiref @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_cchykh5sgr/croots/recipe/asgiref_1661287542702/work
Babel==2.11.0
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
Django @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_eeor5o9jj8/croots/recipe/django_1661292534510/work
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-phonenumber-field==7.0.0
mysqlclient==2.1.1
phonenumbers==8.13.1
Pillow==9.2.0
pytz==2022.6
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
tzdata @ file:///tmp/build/80754af9/python-tzdata_1611693927741/work
wincertstore==0.2
